refactor(api): organize versioned endpoints and router
