+++ python-service/app/api/v1/endpoints/jobspy.py
@@ -6,13 +6,13 @@ from typing import Optional
 from fastapi import APIRouter, HTTPException, Query
 from loguru import logger
 
-from ..models.responses import StandardResponse, create_success_response, create_error_response
-from ..models.jobspy import JobSearchRequest
-from ..services.jobspy_ingestion import get_jobspy_service
-from ..services.queue import get_queue_service
-from ..services.database import get_database_service
-from ..services.scraping import scrape_jobs_async
-from ..services.job_persistence import persist_jobs
+from app.models.responses import StandardResponse, create_success_response, create_error_response
+from app.models.jobspy import JobSearchRequest
+from app.services.jobspy_ingestion import get_jobspy_service
+from app.services.queue import get_queue_service
+from app.services.database import get_database_service
+from app.services.scraping import scrape_jobs_async
+from app.services.job_persistence import persist_jobs
 
 router = APIRouter()
 
@@ -280,7 +280,7 @@ async def jobspy_health_check():
             }
         }
         
-        from ..models.responses import create_success_response
+        from app.models.responses import create_success_response
         return create_success_response(
             data=health_data,
             message="JobSpy service health check completed"
